+++ src/background/index.ts
@@ -1,0 +1,5 @@
+import browser from "webextension-polyfill";
+
+import { handleMessage } from "./messages";
+
+browser.runtime.onMessage.addListener(handleMessage);