+++ src/backgroundScript/index.ts
@@ -1,4 +1,0 @@
-import browser from "webextension-polyfill";
-import handleMessage from "./handleMessage";
-
-browser.runtime.onMessage.addListener(handleMessage);