+++ src/contentScript/index.ts
@@ -1,9 +1,11 @@
-import handleCopyEvent from "./handleCopyEvent";
-import checkWhitelist from "./checkWhitelist";
+import { handleCopyEvent } from "./events";
+import { checkWhitelist } from "./storage";
 
-// Add event listener regardless of whitelisted domain, to ensure it
-// is added as soon as possible. If the current domain is whitelisted,
-// we simply remove the event listener.
+/**
+ * Add event listener regardless of whether or not the domain is whitelisted. This is to make sure
+ * the event listener is added as soon as possible. If the domain is whitelisted, the event
+ * listener will be removed in the checkWhitelist function.
+ */
 document.addEventListener("copy", handleCopyEvent);
 
-checkWhitelist();
+checkWhitelist(window.location.hostname, handleCopyEvent);