+++ src/internal/strings.ts
@@ -1,0 +1,4 @@
+export const removePrefixWWW = (hostname: string) => {
+  if (hostname.startsWith("www.")) return hostname.replace("www.", "");
+  return hostname;
+};