+++ src/options/index.ts
@@ -1,0 +1,3 @@
+import OptionsApp from "./index.svelte";
+
+new OptionsApp({ target: document.body });