+++ src/options/pages/Settings.svelte
@@ -1,0 +1,13 @@
+<script lang="ts">
+  import { Col, Container, Row } from "@sveltestrap/sveltestrap";
+  import SettingsForm from "../components/SettingsForm.svelte";
+</script>
+
+<Container>
+  <Row>
+    <Col>
+      <h1>Settings</h1>
+      <SettingsForm />
+    </Col>
+  </Row>
+</Container>