+++ src/settings/types.ts
@@ -1,9 +1,0 @@
-export type Warning = {
-  type: string;
-  selection: string;
-  clipboard: string;
-};
-
-export type LogState = {
-  [key: string]: Warning[];
-};