+++ src/vite.env.d.ts
@@ -1,0 +1,4 @@
+/// <reference types="svelte" />
+/// <reference types="vite/client" />
+
+declare var __VERSION__: string;