+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2024-02-21
+
+### Added
+
+- Added support for Manifest V3
+
+### Changes
+
+- Moved to Bun
+- Replaced React with Svelte
+- Greatly simplified the way clipboard data is checked for alterations
+  - Instead of comparing text, the extension now simply checks if `clipboardData` of the `ClipboardEvent` contains any data. This only happens when it was altered by Javascript. In all other cases this field is left empty by the browser as a security measure. 
+
 ## [1.3.0] - 2022-04-01
 
 ### Added